% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPIR1threeway.R
\name{FPIR1threeway}
\alias{FPIR1threeway}
\title{Fractional-power interaction regression (FPIR): y ~  x1 + x2 + x3 + x1*x2*x3}
\usage{
FPIR1threeway(y, x1, x2, x3)
}
\arguments{
\item{y}{the dependent variable in regression}

\item{x1}{the first independent variables in regression y ~  x1 + x2 + x3 + x1*x2*x3}

\item{x2}{the second independent variable in regression y ~  x1 + x2 + x3 + x1*x2*x3}

\item{x3}{the third independent variable in regression y ~  x1 + x2 + x3 + x1*x2*x3}
}
\value{
return a data.frame with columns the optimum values of exponents for x1, x2, and x3 in FPIR,
R square value of the whole model, proportion of the variance explained by the interaction term,
coefficients of all terms (the interception, coefficients of x1, x2, x3, and the interaction term).
}
\description{
This function estimates the optimum values of exponents for x1, x2, and x3 in FPIR,
 return model parameters such as regression coefficients, sum of squares of the interaction term, R square, AIC.
}
\examples{

 x1 = runif(100); x2 = runif(100); x3 = runif(100); y = x1 * x2^2.5 * x3^-12 + 0.1*x3; y = jitter(y) # generate a dataset
 out = FPIR1threeway(y, x1, x2, x3)
 out[[2]] # first round exponents
 out[[4]] # second round exponents (tuned results)
 out[[5]] # final model coefficients
 out[[6]] # adjusted R square

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
