% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPIR fractional-power interaction regression
%   ver2.R
\name{FPIR1twoway}
\alias{FPIR1twoway}
\title{Fit fractional-power interaction regression (FPIR) for regression y ~ x1 + x2 + x1^M * x2^N}
\usage{
FPIR1twoway(y, x1, x2)
}
\arguments{
\item{y}{the dependent variable in regression y ~ x1 + x2 + x1^M *x2^N}

\item{x1}{one of the independent variables in regression y ~ x1 + x2 + x1^M *x2^N}

\item{x2}{another independent variable in regression y ~ x1 + x2 + x1^M *x2^N}
}
\value{
return a list including a data.frame with columns of M, N, R square value of the whole model,
proportion of the variance explained by the interaction term, coefficients of all terms
(the interception, x1, x2, and the interaction).
}
\description{
This function estimates the values of exponents (M and N) in Fractional-power interaction regression
(FPIR)，the formula is: y ~ x1 + x2 + x1^M * x2^N.  Return model parameters such as regression coefficients,
sum of squares of the interaction term, and total R squares of the model.
}
\examples{

# Simulated data
x1 = runif(100); x2 = runif(100);
y = x1 * x2^10; y = y + rnorm(length(y), 0, 0.01) # generate a dataset
results = FPIR1twoway (y, x1, x2)
Exp1 = results[[2]]$Ex1; Exp2 = results[[2]]$Ex2
results.2 = FPIR1twowaytune (y, x1, x2, Exp1, Exp2)
results.2[[1]] # parameters for 10000 models
results.2[[2]] # parameters for the best model
results.2[[3]] # regression coefficient
results.2[[4]] # adjusted R square#'

attach(trees)
# First round fitting
results = FPIR1twoway(trees$Volume, trees$Girth, trees$Height) # 1.35, 0.3
# Second round fitting based on the first round fitting
results2 = FPIR1twowaytune(trees$Volume, trees$Girth, trees$Height, 1.35, 0.3)
}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
