% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPIR fractional-power interaction regression
%   ver2.R
\name{FPIR1twowaytune}
\alias{FPIR1twowaytune}
\title{Tune fractional-power interaction regression (FPIR) for regression y ~ x1 + x2 + x1^M * x2^N}
\usage{
FPIR1twowaytune(y, x1, x2, Exp1, Exp2)
}
\arguments{
\item{y}{the dependent variable in regression y ~ x1 + x2 + x1^M *x2^N}

\item{x1}{one of the independent variables in regression y ~ x1 + x2 + x1^M *x2^N}

\item{x2}{another independent variable in regression y ~ x1 + x2 + x1^M *x2^N}

\item{Exp1}{the initial value of M in regression y ~ x1 + x2 + x1^M *x2^N}

\item{Exp2}{the initial value of N in regression y ~ x1 + x2 + x1^M *x2^N}
}
\value{
return a list including tuned values of M and N.
}
\description{
This function tune the estimated values of exponents (M and N) from FPIR1twoway(y, x1, x2).
}
\examples{

# Simulated data
x1 = runif(100); x2 = runif(100);
y = x1 * x2^10; y = y + rnorm(length(y), 0, 0.01) # generate a dataset
results = FPIR1twoway (y, x1, x2)
Exp1 = results[[2]]$Ex1; Exp2 = results[[2]]$Ex2
results.2 = FPIR1twowaytune (y, x1, x2, Exp1, Exp2)
results.2[[1]] # parameters for 10000 models
results.2[[2]] # parameters for the best model
results.2[[3]] # regression coefficient
results.2[[4]] # adjusted R square#'

attach(trees)
# First round fitting
results = FPIR1twoway(trees$Volume, trees$Girth, trees$Height) # 1.35, 0.3
# Second round fitting based on the first round fitting
results2 = FPIR1twowaytune(trees$Volume, trees$Girth, trees$Height, 1.35, 0.3)
}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
