% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPIR fractional-power interaction regression
%   ver2.R
\name{interaction_bayes}
\alias{interaction_bayes}
\title{Fractional-power interaction regression (FPIR) for regression y ~ x1 + x2 + x1^M * x2^N using Bayesian method}
\usage{
interaction_bayes(y, x1, x2, WinBUGS, digits = 3)
}
\arguments{
\item{y}{the dependent variable in regression y ~ x1 + x2 + x1^M *x2^N}

\item{x1}{one of the independent variables in regression y ~ x1 + x2 + x1^M *x2^N}

\item{x2}{another independent variable in regression y ~ x1 + x2 + x1^M *x2^N}

\item{WinBUGS}{the install direction of software WinBUGS}

\item{digits}{controls the number of significant digits to print when printing numeric values (intercept and regression coefficients in formula).}
}
\value{
return a list including a data.frame with estimated model coefficients, and the formula of y ~ x1 + x2 + x1^M * x2^N
}
\description{
This function estimates the values of exponents (M and N) in
Fractional-power interaction regression (FPIR) for regression y ~ x1 + x2 + x1^M * x2^N,
return regression coefficients.
}
\examples{
data(trees)
results = interaction_bayes(trees$Volume, trees$Girth, trees$Height, WinBUGS = "d:/softwares/WinBUGS14/", digits=2)
results

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
