% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPIR1twowaytune.R
\name{FPIR1twowaytune}
\alias{FPIR1twowaytune}
\title{Fractional-power interaction regression (FPIR) for regression y ~ x1 + x2 + x1^M *x2^N}
\usage{
FPIR1twowaytune(y, x1, x2, Exp1, Exp2)
}
\arguments{
\item{y}{the dependent variable in regression}

\item{x1}{one of the independent variables in regression}

\item{x2}{another independent variable in regression}

\item{Exp1}{one of the independent variables in regression}

\item{Exp2}{another independent variable in regression}
}
\value{
return a data.frame with columns of M, N, R square value of the whole model,
proportion of the variance explained by the interaction term, coefficients of all terms
(the interception, x1, x2, and the interaction).
}
\description{
This function estimates the tuned values of exponents (M and N) in FPIR,
 return model parameters such as regression coefficients, sum of squares of the interaction term, R square, AIC.
}
\examples{

 x1 = runif(100); x2 = runif(100);
 y = x1 * x2^10; y = y + rnorm(length(y), 0, 0.01) # generate a dataset
 results = FPIR.1twoway (y, x1, x2)
 Exp1 = results[[2]]$Ex1; Exp2 = results[[2]]$Ex2
 results.2 = FPIR1twowaytune (y, x1, x2, Exp1, Exp2)
 Ex1.tune   = as.numeric(results.2[[2]][1])
 Ex2.tune   = as.numeric(results.2[[2]][2])
 results.2[[2]] # parameters for the best model
 results.2[[3]] # regression coefficients
 results.2[[4]] # adjusted R square
}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
